began modernization
got rid of all the silly stuff i'd done back when i was just learning c++.
there were dynamic allocations that didn't need to happen and algorithms/data structures that already existed in the standard.

only real change is that the notes_t type is no longer an array of possible notes - it's a std::bitset where each bit represents a different note value.
this is to reduce the amount of code, as well as to facilitate a more advanced tuple subgroup elision system i'm planning on implementing.