massive algorithm changes
changed the entire solving algorithm to be less stupid.

previously, we were running through possible values and guessing each time we arrived at an ambiguity.
a couple of problems:
1) this means the recursive steps missed out on all the other tile information, some of which might only have one possible value.
2) this runs the risk of guessing on a tile that e.g. has 9 possible values when there are other, better places to start guessing.

new algorithm goes through all the reduced notes and looks at their counts.
if count > 1, it's put in a priority queue that prioritizes tiles with the least number of possible values.
if count is 1, it's a definite value and is copied to the board.
otherwise there's no possible value, which means the board is illegal.

then, after all that's done, we handle the guesses via the priority queue until we either succeed or run out of notes.
when we're out of notes, we return true iff we succeeded or not.

as far as my test cases go, this reduces test time for each max difficulty board from 20 min (my solution from years upon years ago) to 10.78 milliseconds.
that's pretty good, i'd say.