massive improvements
apparently it's actually faster to not do the tuple logic at all and just take the hit of having to guess more often - go figure.

the tuple logic is now toggled by a preprocessor macro.
when on, behaves as it did before.
when off, performs only tuples of size one (i.e. unique elimination).

added a bunch of test cases.
with the changes above, test cases 3-16 take around .25 milliseconds (big improvement, as before the change they took around 12 milliseconds).
however, test case 1 is about half the speed it used to be (now 5 milliseconds) - after further analysis, it's due to having many more incorrect guesses due to lacking the powerful tuple logic.

but overall it's clear that not doing the tuple logic is much faster in the average case.
after all, all the test cases are the maximum difficulty puzzles i could find.