guess optimization
with the removal of the tuple reduction (to improve avg case use), test 01 was taking about 2x as long as usual.
this commit halfway fixes that by sorting guesses in order of descending likelihood via occurrence counting.

it's not an objectively better solution, as it's a 50/50 change if this lowers or raises the number of incorrect guesses.
i'm not really sure if i'll even leave it in, since while it helps e.g. test 01, there might be some other test case not yet included that takes substantially longer because of this change.
more test cases will be needed to test the better average case.